refactor(character): move CharacterRulesExt to shared/core domain layer
Domain rules (level cap, ability ranges, AC, walk speed) belong in
shared/core alongside the Race and Character domain types, not in the
composeApp presentation layer. Update all import sites accordingly.
