+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/character/presentation/viewmodel/CharacterEditViewModel.kt
@@ -12,12 +12,12 @@ import com.cyrillrx.rpg.character.presentation.CharacterEditState.Loaded
 import com.cyrillrx.rpg.character.presentation.CharacterEditState.Loaded.EditingField
 import com.cyrillrx.rpg.character.presentation.navigation.CharacterRouter
 import com.cyrillrx.rpg.character.presentation.toCharacter
-import com.cyrillrx.rpg.core.presentation.component.dnd.coerceToValidAbilityScore
-import com.cyrillrx.rpg.core.presentation.component.dnd.coerceToValidArmorClass
-import com.cyrillrx.rpg.core.presentation.component.dnd.coerceToValidCharacterLevel
-import com.cyrillrx.rpg.core.presentation.component.dnd.coerceToValidMaxHitPoints
-import com.cyrillrx.rpg.core.presentation.component.dnd.coerceToValidWalkSpeed
-import com.cyrillrx.rpg.core.presentation.component.dnd.defaultWalkSpeed
+import com.cyrillrx.rpg.character.domain.coerceToValidAbilityScore
+import com.cyrillrx.rpg.character.domain.coerceToValidArmorClass
+import com.cyrillrx.rpg.character.domain.coerceToValidCharacterLevel
+import com.cyrillrx.rpg.character.domain.coerceToValidMaxHitPoints
+import com.cyrillrx.rpg.character.domain.coerceToValidWalkSpeed
+import com.cyrillrx.rpg.character.domain.defaultWalkSpeed
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
