--- cmp-ttrpg-companion/composeApp/src/commonTest/kotlin/com/cyrillrx/rpg/core/presentation/component/dnd/CoerceToValidMaxHitPointsTest.kt
@@ -1,20 +1,0 @@
-package com.cyrillrx.rpg.core.presentation.component.dnd
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class CoerceToValidMaxHitPointsTest {
-
-    @Test
-    fun `clamps values below 1 to 1`() {
-        assertEquals(1, 0.coerceToValidMaxHitPoints())
-        assertEquals(1, (-5).coerceToValidMaxHitPoints())
-    }
-
-    @Test
-    fun `returns valid max hit points unchanged`() {
-        assertEquals(1, 1.coerceToValidMaxHitPoints())
-        assertEquals(10, 10.coerceToValidMaxHitPoints())
-        assertEquals(100, 100.coerceToValidMaxHitPoints())
-    }
-}
